% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_2Chain.R
\name{add_2Chain}
\alias{add_2Chain}
\title{Adding two Chain Length:Δ}
\usage{
add_2Chain(vector1, vector2)
}
\arguments{
\item{vector1}{vector(1),like "58:1"}

\item{vector2}{vector(1),like "59:2"}
}
\value{
vector(1)
}
\description{
This function is used to add 2 chain:Δ(it can be 2 vector).Like add "10:1" + "11:1" = "21:2".
}
