% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carbonNumberLaw.R
\name{carbonNumberLaw}
\alias{carbonNumberLaw}
\title{Use carbon number law to filter some wrong Cardiolipin.}
\usage{
carbonNumberLaw(dataframe)
}
\arguments{
\item{dataframe}{dataframe(1) including chain,Δ,rt like (56:0,5).Isomer is allowed.}
}
\value{
dataframe(1) including chain,Δ,rt like (56:0,5) after filtering.
}
\description{
Use carbon number law to filter some wrong Cardiolipin.
}
