% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splicePAFA.R
\name{catchDataframe2Col}
\alias{catchDataframe2Col}
\title{Bind 2 row by column in a dataframe}
\usage{
catchDataframe2Col(row1, row2, dataframe)
}
\arguments{
\item{row1}{numeric(1)}

\item{row2}{numeric(1)}

\item{dataframe}{dataframe(1)}
}
\value{
dataframe(1)
}
\description{
Bind 2 row by column in a dataframe
}
