% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splicePAFA_MLCL.R
\name{catchDataframe3Col}
\alias{catchDataframe3Col}
\title{Bind 3 row by column in a dataframe}
\usage{
catchDataframe3Col(row1, row2, row3, dataframe)
}
\arguments{
\item{row1}{numeric(1)}

\item{row2}{numeric(2)}

\item{row3}{numeric(3)}

\item{dataframe}{dataframe(1)}

\item{row4}{numeric(4)}
}
\value{
dataframe(1)
}
\description{
Bind 3 row by column in a dataframe
}
