% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splicePAFA.R
\name{catchDataframe4Col}
\alias{catchDataframe4Col}
\title{Bind 4 row by column in a dataframe}
\usage{
catchDataframe4Col(row1, row2, row3, row4, dataframe)
}
\arguments{
\item{row1}{numeric(1)}

\item{row2}{numeric(2)}

\item{row3}{numeric(3)}

\item{row4}{numeric(4)}

\item{dataframe}{dataframe(1)}
}
\value{
dataframe(1)
}
\description{
Bind 4 row by column in a dataframe
}
