% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkMS2.R
\name{catchMS2}
\alias{catchMS2}
\title{Catch MS2}
\usage{
catchMS2(object, MS2_name)
}
\arguments{
\item{object}{a findmyCL object}

\item{MS2_name}{character(1)}
}
\value{
Spectrum2
}
\description{
Catch MS2
}
