% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matchMS2.R
\name{check_HeadFAPA}
\alias{check_HeadFAPA}
\title{Checking PA and FA for a MS2}
\usage{
check_HeadFAPA(MS1, MS2, ppm = 30, ms2checkresult_name)
}
\arguments{
\item{MS1}{list(1)}

\item{MS2}{list(1)}

\item{ppm}{numeric(1)}
}
\description{
Checking if a MS2 have its corresponding PA or FA and splice them,use \code{\link[=findPAFA]{findPAFA()}} to find PA and FA.
}
\seealso{
\code{\link[=splicePAFA]{splicePAFA()}}
}
