% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matchMS2.R
\name{check_MS1_MS2_HeadFAPA}
\alias{check_MS1_MS2_HeadFAPA}
\title{Checking if a big MS1 have its corresponding MS2}
\usage{
check_MS1_MS2_HeadFAPA(MS1, ppm = 30, ms2checkresult_name)
}
\arguments{
\item{MS1}{list(1)}

\item{ppm}{numeric(1)}

\item{pb}{1}
}
\description{
checking if a aggregation which have MS1 and MS2 have PA or FA and splice PA and FA in cardiolipin.
}
