% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkMS2.R
\name{checkMS2_main}
\alias{checkMS2_main}
\title{Check every MS1 with its MS2 in a MS1-result}
\usage{
checkMS2_main(object, ppm = 5, ms1matchresult_name = "CL")
}
\arguments{
\item{object}{a findmyCL object}

\item{ppm}{numeric(1) defining the maximal tolerated m/z deviation in consecutive scans in parts per million (ppm) for the initial ROI definition when matching with the precursorMz of MS2,default is 5.}

\item{ms1matchresult_name}{character(1)}
}
\value{
list(1)
}
\description{
Check every MS1 with its MS2 in a MS1-result
}
