% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkMS2.R
\name{checknoMS2_main}
\alias{checknoMS2_main}
\title{Check which MS1 has no MS2 in every MS1-match-result}
\usage{
checknoMS2_main(object, ppm = 5, ms1matchresult_name = "CL")
}
\arguments{
\item{object}{a findmyCL object}

\item{ppm}{numeric(1) defining the maximal tolerated m/z deviation in consecutive scans in parts per million (ppm) for the initial ROI definition when matching with the precursorMz of MS2,default is 5.}

\item{ms1matchresult_name}{character(1)}
}
\value{
list(1)
}
\description{
Check which MS1 has no MS2 in every MS1-match-result
}
