% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtFilter.R
\name{choose_list}
\alias{choose_list}
\title{Choose 1-MS1 vs 1 Cardiolipin and 1-MS1 vs more Cardiolipin.}
\usage{
choose_list(list, mode = 1)
}
\arguments{
\item{list}{list(1),a bigMS1 contain MS1,MS2,splice-Cardiolipin-result.}

\item{mode}{numeric(1),modes used to select 1 vs 1 or 1 vs more than 1.}
}
\value{
list(1)
}
\description{
Choose 1-MS1 vs 1 Cardiolipin and 1-MS1 vs more Cardiolipin.
}
