% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createFrameByVector.R
\name{createFrameByVector}
\alias{createFrameByVector}
\title{Creating a dataframe by vector}
\usage{
createFrameByVector(vector, n)
}
\arguments{
\item{vector}{vector(1)}

\item{n}{numeric(1)}
}
\value{
dataframe(1)
}
\description{
Decompose a vector.Every n elements in this vector constitute a row in the new dataframe.
}
