% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadCentroidData.R
\name{cutMS2}
\alias{cutMS2}
\title{Delete the peak with the lowest intensity in all MS2.}
\usage{
cutMS2(MS2, MS2cutRadio = 0.1)
}
\arguments{
\item{MS2}{environment(1)}

\item{MS2cutRadio}{numeric(1),default is 10\%.}
}
\value{
environment(1)
}
\description{
Delete the peak with the lowest intensity in all MS2.
}
\seealso{
\code{\link[=cutMS2_main]{cutMS2_main()}}
}
