% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadCentroidData.R
\name{cutMS2_main}
\alias{cutMS2_main}
\title{Delete the peak with the lowest intensity in a MS2.}
\usage{
cutMS2_main(siMS2, MS2cutRadio = 0.1)
}
\arguments{
\item{siMS2}{list(1)}

\item{MS2cutRadio}{numeric(1),default is 10\%.}
}
\value{
list(1)
}
\description{
Delete the peak with the lowest intensity in a MS2.
}
