% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defineMyClass.R
\name{defineMyClass}
\alias{defineMyClass}
\title{A function to define how the data store in the findmyCL object}
\usage{
defineMyClass()
}
\description{
A function to define how the data store in the findmyCL object
}
