% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtFilter.R
\name{delete_redundant_PA}
\alias{delete_redundant_PA}
\title{Delete redundant splicePA-result.}
\usage{
delete_redundant_PA(MS2, chain)
}
\arguments{
\item{MS2}{list(1)}

\item{chain}{vector(1) or character(1)}
}
\value{
list(1)
}
\description{
Use Chain Length:Δ in spliceFA-result(it exist mean a Cardiolipin can be spliced by PA which also can be spliced by FA).But some splicePA-result can not be spliced by FA(it has not be deleted yet).So use this function to delete these splicePA-result.Work for \code{\link[=delete_no_FA_MS1]{delete_no_FA_MS1()}}
}
