% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtFilter.R
\name{delete_redundant_PA_main}
\alias{delete_redundant_PA_main}
\title{Delete redundant splicePA-result.}
\usage{
delete_redundant_PA_main(splicePA_result, chain)
}
\arguments{
\item{splicePA_result}{list(1),contain splicePA,Chain Length:Δ,oxygen}

\item{chain}{vector(1)}
}
\value{
list(1)
}
\description{
Work for \code{\link[=delete_redundant_PA]{delete_redundant_PA()}}
}
