% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deleteCardiolipinMoreThanOne.R
\name{deleteCardiolipinMoreThanOne}
\alias{deleteCardiolipinMoreThanOne}
\title{Delete redundant Cardiolipin}
\usage{
deleteCardiolipinMoreThanOne(object)
}
\arguments{
\item{object}{a findmyCL object}
}
\value{
a findmyCL object
}
\description{
Some MS1 have more than 1 corresponding Cardiolipin,but it should only one  Cardiolipin in a MS1,this function use seed which is actually a Cardiolipin get only one Cardiolipin named rtseed to choose which Cardiolipin is the right Cardiolipin in MS1 which get more than 2 Cardiolipin.Use retention time rule to filter some chain:Δ.The longer the chain,the longer the retention time.The greater the unsaturation,The shorter the retention time.
}
