% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deleteCardiolipinMoreThanOne.R
\name{deleteCardiolipinMoreThanOne_main}
\alias{deleteCardiolipinMoreThanOne_main}
\title{Delete redundant Cardiolipin for a single MS2matchresult}
\usage{
deleteCardiolipinMoreThanOne_main(object, ms2matchresult_name = "CL")
}
\arguments{
\item{object}{a findmyCL object}

\item{ms2matchresult_name}{character(1)}
}
\value{
list(1)
}
\description{
Delete redundant Cardiolipin for a single MS2matchresult
}
