% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fappend.R
\name{fappend}
\alias{fappend}
\title{Append all elements in list2 to list1.}
\usage{
fappend(list1, list2)
}
\arguments{
\item{list1}{the list appended}

\item{list2}{the list you choose to append}
}
\value{
list(1)
}
\description{
A improvement of \code{\link[=append]{append()}}.If you use test <- append(x = list1 , values = list2).The result is the list2 is added to the list1.But what we want is add whole elements in list2 to list1 rather than adding a list2 to list1.So this function
}
