% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtFilter.R
\name{get_all_chain}
\alias{get_all_chain}
\title{Get all Chain Length:Δ in a MS2.}
\usage{
get_all_chain(MS2)
}
\arguments{
\item{MS2}{list(1)}
}
\value{
matrix(1)
}
\description{
Extract all Chain Length:Δ from splicePA-result && spliceFA-result in a MS2.
}
