% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtFilter.R
\name{get_chain}
\alias{get_chain}
\title{Get all Chain Length:Δ in a list.}
\usage{
get_chain(list)
}
\arguments{
\item{list}{list(1),contain splice-result and Chain Length:Δ.}
}
\value{
character(1)
}
\description{
Extract a Chain Length:Δ in a splice-result(a list).
}
\seealso{
\code{\link[=get_all_chain]{get_all_chain()}}
}
