% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deleteCardiolipinMoreThanOne.R
\name{getAllChainRt}
\alias{getAllChainRt}
\title{Get all rt,chain,Δ in a bigMS1.}
\usage{
getAllChainRt(bigMS1)
}
\arguments{
\item{bigMS1}{list(1)}
}
\value{
dataframe(1)
}
\description{
Get all rt,chain,Δ in a bigMS1.A bigMS1 is consist of MS1 and some MS2.Take rt,chain,Δ
}
