% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadCentroidData.R
\name{loadCentroidData}
\alias{loadCentroidData}
\title{Load rawdata in centroid mode}
\usage{
loadCentroidData(
  file,
  ppm = 5,
  peakwidth = c(5, 30),
  prefilter = c(4, 5000),
  snthresh = 3,
  mzdiff = -0.001,
  noise = 0,
  integrate = 1L,
  fitgauss = FALSE,
  MS2cutRadio = 0.1,
  rtcut = 6
)
}
\arguments{
\item{file}{a mzml file in centroid mode}

\item{ppm}{numeric(1) defining the maximal tolerated m/z deviation in consecutive scans in parts per million (ppm) for the initial ROI definition.Default is 5.}

\item{peakwidth}{numeric(2) with the expected approximate peak width in chromatographic space. Given as a range (min, max) in seconds.Default is c(5,30).}

\item{prefilter}{numeric(2): c(k, I) specifying the prefilter step for the first analysis step (ROI detection). Mass traces are only retained if they contain at least k peaks with intensity >= I.Default is c(4,5000).}

\item{snthresh}{numeric(1) defining the signal to noise ratio cutoff.Default is 3.}

\item{mzdiff}{numeric(1) representing the minimum difference in m/z dimension required for peaks with overlapping retention times; can be negative to allow overlap. During peak post-processing, peaks defined to be overlapping are reduced to the one peak with the largest signal.Default is -0.001.}

\item{noise}{numeric(1) allowing to set a minimum intensity required for centroids to be considered in the first analysis step (centroids with intensity < noise are omitted from ROI detection).Default is 0.}

\item{integrate}{Integration method. For integrate = 1 peak limits are found through descent on the mexican hat filtered data, for integrate = 2 the descent is done on the real data. The latter method is more accurate but prone to noise, while the former is more robust, but less exact.Default is 1.}

\item{fitgauss}{logical(1) whether or not a Gaussian should be fitted to each peak. This affects mostly the retention time position of the peak.Default is FALSE.}

\item{MS2cutRadio}{numeric(1) cut the last intensity in MS2}

\item{rtcut}{numeric(1) half peak width to cut the MS2}
}
\value{
a findmyCL object
}
\description{
A function use function xcmsSet of  package "xcms" and "MSnbase" for deceting the peak of mzml file in centroid mode.We set some defaults in this function like ppm and snthresh.If your data format is not centroid,you can use \code{\link[=profilePretreat]{profilePretreat()}}
}
\seealso{
\code{\link[=profilePretreat]{profilePretreat()}} turn a mzml file in profile mode to  centroid mode for the input of this function
}
