% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matchDatabase.R
\name{matchDatabase}
\alias{matchDatabase}
\title{Matching dataframe with a database}
\usage{
matchDatabase(vector, vector_column, ppm = 5, database, database_column, pb)
}
\arguments{
\item{vector}{vector(1)}

\item{vector_column}{numeric(1) which column you want to compare}

\item{ppm}{The floating value you can tolerate most}

\item{database}{dataframe(1)}

\item{database_column}{numeric(1) which column in database you choose to compare}
}
\value{
a dataframe
}
\description{
Matching dataframe with a database consist of list.This function is used in \code{\link[=matchMS1]{matchMS1()}}.
}
