% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matchDatabaseMS2.R
\name{matchDatabaseMS2}
\alias{matchDatabaseMS2}
\title{Matching dataframe with a database
#' @description Matching dataframe with a database consist of list.This function is used in \code{\link[=matchMS2]{matchMS2()}}.}
\usage{
matchDatabaseMS2(vector, vector_column, ppm = 5, database, database_column)
}
\arguments{
\item{vector}{vector(1)}

\item{vector_column}{numeric(1) which column you want to compare}

\item{ppm}{The floating value you can tolerate most}

\item{database}{dataframe(1)}

\item{database_column}{numeric(1) which column in database you choose to compare}
}
\value{
a dataframe
}
\description{
Matching dataframe with a database
#' @description Matching dataframe with a database consist of list.This function is used in \code{\link[=matchMS2]{matchMS2()}}.
}
