% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matchMS1.R
\name{matchMS1_main}
\alias{matchMS1_main}
\title{Match MS1 with the CL-database.}
\usage{
matchMS1_main(object, ppm = 5, database_name = "CL")
}
\arguments{
\item{object}{a findmyCL object.}

\item{ppm}{numeric(1),default is 5.}

\item{database_name}{character(1)}
}
\value{
dataframe(1)
}
\description{
Match MS1 with the CL-database.
}
\seealso{
\code{\link[=matchDatabase]{matchDatabase()}}
}
