% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matchMS2.R
\name{matchMS2}
\alias{matchMS2}
\title{Matching the MS2 database}
\usage{
matchMS2(object, ppm = 30, workers = 2)
}
\arguments{
\item{object}{a findCL object after \code{\link[=checkMS2]{checkMS2()}}}

\item{ppm}{numeric(1) defining the maximal tolerated m/z deviation in consecutive scans in parts per million (ppm) for the initial ROI definition when match our database.Default is 30.}

\item{workers}{numeric(1) Number of multithreaded cores you use in this function.Default is 2.}
}
\value{

}
\description{
This function use result in MS2-checkresult to search PA&FA and splice Cardiolipin.The MS2 database consists of PA&FA database.Before searching PA&FA,we use  \code{\link[=check_MS1_MS2_HeadFAPA]{check_MS1_MS2_HeadFAPA()}} to check if a MS2 have a head group(mz:152.9963485799).If this MS2 has not head group,we would delete it.If this MS2 have head group,then we would find PA&FA in this MS2.After searching PA&FA,we would splice PA&FA into Cardiolipin \code{\link[=splicePAFA]{splicePAFA()}} .If this MS2 cannot splice into Cardiolipin,we would delete this MS2.Moreover,we would delete a MS1 and its MS2 if we connot find Cardiolipin in all its MS2.
}
