% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matchMS2.R
\name{matchMS2_main}
\alias{matchMS2_main}
\title{Search every MS2-checking for their PA&FA and splice them.}
\usage{
matchMS2_main(object, ppm = 30, ms2checkresult_name = "CL", workers = 2)
}
\arguments{
\item{object}{findmyCLclass(1)}

\item{ppm}{numeric(1)}

\item{ms2checkresult_name}{character(1)}
}
\description{
search every MS2-checking for their PA&FA and splice them,every element use \code{\link[=check_MS1_MS2_HeadFAPA]{check_MS1_MS2_HeadFAPA()}} to search and splice.
}
