% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profilePretreat.R
\name{profilePretreat}
\alias{profilePretreat}
\title{Data preprocessing}
\usage{
profilePretreat(
  file,
  smooth_method = "MovingAverage",
  halfWindowSize = 2,
  polynomialOrder = 3,
  pickPeaks_method = "kNeighbors",
  kNeighbors_par = 1,
  signalPercentage = 50,
  output_file
)
}
\arguments{
\item{file}{character(1)  mzml file in profile mode}

\item{smooth_method}{character(1) smoothing method for your data.Like MovingAverage and SavitzkyGolay.}

\item{halfWindowSize}{numeric(1) if you choose MovingAverage for your smooth,you have to set the halfWindowSize,default is 2.}

\item{polynomialOrder}{character(1) if you choose SavitzkyGolay for your smooth,you have to set the polynomialOrder,default is 3.}

\item{pickPeaks_method}{character(1) peak picking method for your data.Like kNeighbors and descendPeak.}

\item{kNeighbors_par}{numeric(1) if you choose kNeighbors for your peak picking method,you have to set the parameter,kNeighbors_par,default is 1.}

\item{signalPercentage}{numeric(1) if you choose descendPeak for your peak picking method,you have to set the parameter,signalPercentage,default is 50.}

\item{output_file}{character(1) the name of your output mzml file.}
}
\value{
a mzml file
}
\description{
A function to protreat the mzml file in profile mode into a mzml file in centroid mode,the result of this function is to output a mzml file in centroid mode.https://bioconductor.riken.jp/packages/3.7/bioc/vignettes/MSnbase/inst/doc/MSnbase-centroiding.html
}
\examples{
profilePretreat(file = "QC-NGE-ALL.mzML" , smooth_method = "MovingAverage" , halfWindowSize = 2 , pickPeaks_method = "kNeighbors" , kNeighbors_par = 1 , output_file = "Test.mzml")
}
