% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbindList.R
\name{rbindList}
\alias{rbindList}
\title{Bind all dataframe in a list with repetiteve colnames.}
\usage{
rbindList(list)
}
\arguments{
\item{list}{list(1)}
}
\value{
dataframe(1)
}
\description{
Bind all dataframe in a list with repetiteve colnames.
}
