% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbindList.R
\name{rbindList_main}
\alias{rbindList_main}
\title{Change every colnames of dataframe in the list.}
\usage{
rbindList_main(dataframe)
}
\arguments{
\item{dataframe}{dataframe(1)}
}
\value{
dataframe(1)
}
\description{
Change every colnames of dataframe in the list.
}
