% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtFilter.R
\name{rtfilter}
\alias{rtfilter}
\title{Split every MS2matchresult into half}
\usage{
rtfilter(object)
}
\arguments{
\item{object}{a findmyCL class}
}
\value{
a findmyCL class
}
\description{
This is a function to split MS2matchresult into half.Half have only one cardiolipin,the other have more than tow match cardiolipin.The former is used to select which is the right cardiolipin in the latter half.
}
