% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtFilter.R
\name{rtfilter_main}
\alias{rtfilter_main}
\title{Split a MS2match&splice result into half}
\usage{
rtfilter_main(object, deal_name = "CL")
}
\arguments{
\item{object}{a findmyCL class}

\item{deal_name}{character(1) the name of MS2matchresult,like "CL"}
}
\value{
list(1)
}
\description{
Delete the redundant Cardiolipin in MS1(a MS1 may have more than 1 possible Cardiolipin and it may can not be spliced into Cardiolipin).Delete redundant splicePA result in every MS2(some splicePA result may can not be spliced by its FA in its MS2).Work for \code{\link[=rtfilter]{rtfilter()}}
}
