% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carbonNumberLaw.R
\name{score}
\alias{score}
\title{Scoring a group of Cardiolipin.}
\usage{
score(dataframe)
}
\arguments{
\item{dataframe}{dataframe(1) including chain,Δ,rt.Its (chain,Δ) can not repeat beside itself.}
}
\value{
dataframe(1) including chain,Δ,rt and their score
}
\description{
Scoring a group of Cardiolipin.
}
\seealso{
\code{\link[=score_main]{score_main()}}
}
