% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carbonNumberLaw.R
\name{score_main}
\alias{score_main}
\title{Scoring a single Cardiolipin.Its (chain,Δ) can not repeat beside itself.}
\usage{
score_main(x, dataframe)
}
\arguments{
\item{x}{dataframe(1) the single Cardiolipin}

\item{dataframe}{dataframe(1) including all Cardiolipin}
}
\value{
numeric(1) the score
}
\description{
Scoring a single Cardiolipin.Its (chain,Δ) can not repeat beside itself.
}
