% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkMS2.R
\name{searchMS2}
\alias{searchMS2}
\title{Search which MS1 have its MS2}
\usage{
searchMS2(
  object,
  vector,
  vector_mz_column,
  vector_rtmin_column,
  vector_rtmax_column,
  ppm = 5,
  MS2,
  MS2_column,
  pb
)
}
\arguments{
\item{object}{a findmyCL object}

\item{vector}{vector(1)}

\item{vector_mz_column}{numeric(1)}

\item{vector_rtmin_column}{numeric(1)}

\item{vector_rtmax_column}{numeric(1)}

\item{ppm}{numeric(1),defining the maximal tolerated m/z deviation in consecutive scans in parts per million (ppm) for the initial ROI definition when matching with the precursorMz of MS2,default is 5.}

\item{MS2}{dataframe(1)}

\item{MS2_column}{numeric(1)}

\item{pb}{process bar}
}
\value{
list(1)
}
\description{
Search which MS1 have its MS2
}
