% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splicePAFA_MLCL.R
\name{splice1FA_1PA_into_MLCL}
\alias{splice1FA_1PA_into_MLCL}
\title{Splice 1FA and 1PA into a MLCL}
\usage{
splice1FA_1PA_into_MLCL(spliceFA_result_dataframe, FA, chain_double, oxygen)
}
\arguments{
\item{spliceFA_result_dataframe}{dataframe(1),have 1 row}

\item{FA}{dataframe(1)}

\item{chain_double}{character(1),Chain Length:Δ from MS1}

\item{oxygen}{numeric(1),Oxform from MS1}
}
\value{
dataframe(1)
}
\description{
Give a PA which is spliced by 2FA.Use this PA to search which FA in FA(dataframe) can be splice into a MLCL together.
}
