% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splicePAFA.R
\name{splice2FA}
\alias{splice2FA}
\title{Splice 2 FA from the result in PA-splicing}
\usage{
splice2FA(splicePA, MS2)
}
\arguments{
\item{splicePA}{list(1)}

\item{MS2}{list(1)}
}
\description{
Splice 2 FA from the result in PA-splicing
}
