% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splicePAFA_DLCL.R
\name{splice2FA_DLCL}
\alias{splice2FA_DLCL}
\title{Splice 2FA into a DLCL}
\usage{
splice2FA_DLCL(FA, chain_double, oxygen)
}
\arguments{
\item{FA}{dataframe(1)}

\item{chain_double}{character(1)，like "14:0".}

\item{oxygen}{numeric(1)}
}
\value{
list(1)
}
\description{
Splice 2FA into a DLCL which have no PA.
}
