% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splicePAFA.R
\name{splice2FA_main}
\alias{splice2FA_main}
\title{Splice 2 FA from the result in PA-splicing}
\usage{
splice2FA_main(splicePA_dataframe, MS2)
}
\arguments{
\item{splicePA_dataframe}{dataframe(1)}

\item{MS2}{list(1)}
}
\description{
Splice 2 FA from the result in PA-splicing
}
