% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splicePAFA_MLCL.R
\name{splice2FA_MLCL}
\alias{splice2FA_MLCL}
\title{Splice 1PA & 1FA into MLCL.}
\usage{
splice2FA_MLCL(splicePA, FA)
}
\arguments{
\item{splicePA}{list(1),including PA,Chain Length:Δ,oxygen.}

\item{FA}{dataframe(1)}
}
\value{
list(1)
}
\description{
First,splice 2FA into 1PA.Second,search the third FA which can be splice into a MLCL with this 2FA(can be splice into 1PA).
}
