% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splicePAFA_MLCL.R
\name{splice2FA_MLCL_main}
\alias{splice2FA_MLCL_main}
\title{Splice 2FA into a PA in MLCL}
\usage{
splice2FA_MLCL_main(splicePA_dataframe, FA)
}
\arguments{
\item{splicePA_dataframe}{dataframe(1),have 1 row.}

\item{FA}{dataframe(1)}
}
\value{
dataframe(1)
}
\description{
Use \code{\link[=splice2PA]{splice2PA()}} to splice 2FA into a PA,return a dataframe.
}
