% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splicePAFA.R
\name{splice2FA_vector}
\alias{splice2FA_vector}
\title{Splice 2 FA from the result in PA-splicing}
\usage{
splice2FA_vector(vector, MS2)
}
\arguments{
\item{vector}{vector(1)}

\item{MS2}{list(1)}
}
\description{
Splice 2 FA from the result in PA-splicing
}
