% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splicePAFA.R
\name{splice2PA}
\alias{splice2PA}
\title{Splice 2PA into a cardiolipin}
\usage{
splice2PA(PA, chain_double, oxygen)
}
\arguments{
\item{PA}{dataframe(1)}

\item{chain_double}{character(1)}

\item{oxygen}{numeric(1)}
}
\value{
list(1),including splicePA-result,its corresponding Chain Length:Δ && oxygen
}
\description{
Splice 2PA into a cardiolipin
}
