% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splicePAFA_MLCL.R
\name{splice3FA_MLCL}
\alias{splice3FA_MLCL}
\title{Splice 3FA into a MLCL}
\usage{
splice3FA_MLCL(FA, chain_double, oxygen)
}
\arguments{
\item{FA}{dataframe(1)}

\item{chain_double}{character(1)，like "14:0".}

\item{oxygen}{numeric(1)}
}
\value{
list(1)
}
\description{
Splice 3FA into a MLCL which have no PA.
}
