% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splicePAFA.R
\name{splice4FA}
\alias{splice4FA}
\title{Splice 4 FA into a cardiolipin}
\usage{
splice4FA(FA, chain_double, oxygen)
}
\arguments{
\item{FA}{dataframe(1)}

\item{chain_double}{character(1)}

\item{oxygen}{numeric(1)}
}
\value{
list(1)
}
\description{
Splice 4 FA into a cardiolipin
}
