% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splicePAFA_DLCL.R
\name{splicePA2FA_DLCL}
\alias{splicePA2FA_DLCL}
\title{Splice 2FA into a PA.}
\usage{
splicePA2FA_DLCL(splicePA, FA)
}
\arguments{
\item{splicePA}{list(1),including PA,Chain Length:Δ,oxygen.}

\item{FA}{dataframe(1)}
}
\value{
list(1)
}
\description{
In DLCL,we find a PA first.Then,search which FA can compose this PA.
}
