% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splicePAFA_DLCL.R
\name{splicePA_DLCL}
\alias{splicePA_DLCL}
\title{Search which PA can be a DLCL.}
\usage{
splicePA_DLCL(PA, chain_double, oxygen)
}
\arguments{
\item{PA}{dataframe(1)}

\item{chain_double}{character(1)}

\item{oxygen}{numeric(1)}
}
\value{
list(1),including splicePA-result,its corresponding Chain Length:Δ && oxygen
}
\description{
A PA can be a DLCL if its Chain Length:Δ and oxform can match this DLCL.
}
