% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splicePAFA_MLCL.R
\name{splicePA_MLCL}
\alias{splicePA_MLCL}
\title{Search which PA can splice into a MLCL with a FA.}
\usage{
splicePA_MLCL(PA, chain_double, oxygen, FA)
}
\arguments{
\item{PA}{dataframe(1)}

\item{chain_double}{character(1)}

\item{oxygen}{numeric(1)}
}
\value{
list(1),including splicePA-result,its corresponding Chain Length:Δ && oxygen
}
\description{
Search which PA can splice into a MLCL with a FA.
}
