% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splicePAFA.R
\name{splicePAFA}
\alias{splicePAFA}
\title{Splice PA and FA from a MS2 and MS1 in CL}
\usage{
splicePAFA(MS2, MS1)
}
\arguments{
\item{MS2}{list(1),including MS2 information,PA&FA search result}

\item{MS1}{dataframe(1),including mz,rt,rtmin,rtmax ...}
}
\value{
list(1),a new MS2,adding PA&FA splicing result
}
\description{
A function to search PA and FA from MS2.If this MS1 can not search its 2 PA,delete this MS1(this MS1 may not have just one corresponding Cardiolipin).If this MS2 can search its PA but can not searth corresponding 2 FA or it can just match one FA,delete this MS2.
}
