% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splicePAFA_DLCL.R
\name{splicePAFA_DLCL}
\alias{splicePAFA_DLCL}
\title{Splice PA and FA from a MS2 and MS1 in DLCL}
\usage{
splicePAFA_DLCL(MS2, MS1)
}
\arguments{
\item{MS2}{list(1),including MS2 information,PA&FA search result}

\item{MS1}{dataframe(1),including mz,rt,rtmin,rtmax ...}
}
\value{
list(1),a new MS2,adding PA&FA splicing result
}
\description{
DLCL:Double hydrolyzed cardiolipin(have 1 PA).A function to search PA and FA from MS2.If this MS1 can not search its 2 PA,delete this MS1(this MS1 may not have just one corresponding Cardiolipin).If this MS2 can search its 2 PA but can not search corresponding 2 FA or it can just match one FA,delete this MS2.
}
