% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/turnDataframeList.R
\name{turnDataframeList}
\alias{turnDataframeList}
\title{Turn a dataframe into list by row}
\usage{
turnDataframeList(dataframe)
}
\arguments{
\item{dataframe}{dataframe(1)}
}
\value{
list(1)
}
\description{
Turn a dataframe into list by row
}
