% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/turnVectorDataframe.R
\name{turnVectorDataframe}
\alias{turnVectorDataframe}
\title{Turn a vector to a dataframe}
\usage{
turnVectorDataframe(vector)
}
\arguments{
\item{vector}{vector(1)}
}
\value{
dataframe(1)
}
\description{
A MS1 may has only 1 corresponding Cardiolipin.But some MS1 may has more than 1 corresponding Cardiolipins.So this function is used to turn these 1 vs much Cardiolipin(vector) into a dataframe.
}
